examples/requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
html2text>=2024.2.26
